[JARS-119] do not assume there is exactly one resource with uri
just take the first one
